fixed pre-commit failures

+++ inference-engine/tests/functional/inference_engine/transformations/convert_gather_0d_test.cpp
@@ -25,7 +25,7 @@ TEST(TransformationTests, ConvertGather0DStatic1) {
     std::shared_ptr<Function> f(nullptr), f_ref(nullptr);
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{15, 4, 20, 28});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{15, 4, 20, 28});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset7::Gather>(input, indices, axis_const);
 
@@ -56,7 +56,7 @@ TEST(TransformationTests, ConvertGather0DStatic2) {
     std::shared_ptr<Function> f(nullptr), f_ref(nullptr);
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset7::Gather>(input, indices, axis_const);
